# optimize unaligned operator same as arm.