@huangwenjun06 2018-08-20
optimize code on mips platform  submit finish. Could you help to review it?